Introduce Webpack
This does not represent any functional change to the behavior of the library. Since this component is getting reasonably complex, this diff adds bundling of files via Webpack. I have broken off a first small piece into geometry.js just as an example, and we should continue to split things up further.

Notes:
- Our webpack config emits a umd compatible module. Currently in master we only support es6 style modules.
- See the notes in ./copy-to-example.sh to see how we are using the dist of this lib in the example app
- Some useful resources: https://webpack.js.org/guides/author-libraries/ and https://github.com/kalcifer/webpack-library-example