Support devices that have both touch screen and pointer inputs. (#25)
Fixes #14

Some devices have both a touch screen as well as a pointer input, like a mouse. In master, we only attach events for one of those when the component loads (based on whether or not the device supports touch events). For devices supporting both, this means that the touch event takes priority so that the mouse events do not get attached. To fix this, this diff indiscriminately attaches all events.

Warning -- I have not yet tested on such a device and would like to do so before landing. I have tested on a desktop, a mobile device, and a mobile simulator so far.