This diff fixes 2 issues, both were bugs related to dragging.
The first change an issue where the map was changing translation when dragging outside of container.
This bug can be reproduced by:
- Perform a normal drag inside of the container, then
- Drag somewhere outside of the container, which should have no impact on the translation of the contents, however you will see that the contents will change translation.
This was happening because the touchEnd event handler of the initial drag was not being called because we were improperly doing a stopPropagation that blocked it. Now, the touchEnd event is attached at the window and triggered during the 'capture' phase of the event, ensuring that it will get called *before* any other handlers have a chance to stop prop.

The second change fixes #20.
The change here was simply to make startPointerInfo undefined whenever `pointers` was empty, instead of having startPointerInfo defined with an empty pointers list.

Also fixes #28 by adding a changelog.